se cambio el double a bigdecimal

+++ src/main/java/shareddtos/billmodule/BillItem/BillItemDTO.java
@@ -1,5 +1,7 @@
 package shareddtos.billmodule.BillItem;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,6 +14,6 @@ public class BillItemDTO {
     private Long productId;
     private String name;
     private int requestedQuantity;
-    private Double price;
-    private Double subTotal;
+    private BigDecimal price;
+    private BigDecimal subTotal;
 }
