+++ src/main/java/shareddtos/billmodule/BillItem/CreateBillItemDTO.java
@@ -1,5 +1,7 @@
 package shareddtos.billmodule.BillItem;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,7 +14,7 @@ public class CreateBillItemDTO {
     private Long productId;
     private String name;
     private int requestedQuantity;
-    private Double price;
-    private Double subTotal;
+    private BigDecimal price;
+    private BigDecimal subTotal;
 
 }
